removed test files, edited other docs
