+++ package.json
@@ -29,19 +29,14 @@
     "bcrypt": "^0.8.6",
     "body-parser": "^1.15.0",
     "express": "^4.13.4",
-    "gulp": "^3.9.1",
-    "jsonwebtoken": "^6.2.0",
     "mongodb": "^2.1.16",
     "mongoose": "^4.4.13"
   },
   "devDependencies": {
     "angular": "^1.5.5",
-    "chai": "^3.5.0",
-    "chai-http": "^2.0.1",
+    "gulp": "^3.9.1",
     "gulp-eslint": "^2.0.0",
-    "gulp-mocha": "^2.2.0",
     "gulp-protractor": "^2.3.0",
-    "mocha": "^2.4.5",
     "webpack-stream": "^3.2.0"
   }
 }
